initial commit
ipip intelligent pip installer. No more guessing is the goal hahaha. We now accept assumptions : )

pip is cool, but we need to evolve the package world

Added command functionality in order to perform basic file operations using the ai. This includes memory of previously moved/interacted files in order to perform continued operations via prompting.

This is a package installer, so this is a bit much hahahahaha. Oh well.

Please see docs for some functionality details.